Refactored musicUtils to multiple files
